Add files via upload
Career Decisions is used to best determine which career best suits you through evaluation of different factors in a scale of 1-5, 5 being the best factor favoring you. The total score determines the best career you suit in. Can be used for career coaches and new students who are yet to find their careers and what they are best suited in.
Crystalpools sales report gives a sales report for sales of different items, region, sales person and is used to determine the profit and commission earned by each sales person. The pivot table and pie chart are used to show the sales for each sales person and determining the best sales person. This can be used in appraising and in forecasting the future of sales in the company to know what best suits which location and for which sales person.
Gradebook project is used to evaluate employees of  a company and analyze their performance. Employees are  ranked depending on different skill sets. It is used to determine areas of weakness of different employees, areas of their strength and those who are ardent. Also can be used to help the HR know whom to Fire or appraise.
Payroll project is a payroll used by the HR to record hours each employee worked each week, the hourly pay. It is used to determine the overtime worked, weekly pay for each employee that is the overtime pay and weekly payments. Used to calculate the Monthly salary for each employee. If improved it can be used to forecast the rate at which the payroll will grow and what should be the adjustments to the payroll.